Fix integer overflow in the return of `RouteFinder::find_next_hop`
Previously, `RouteFinder::find_next_hop` has a return type of `uint16_t`.
This resulted in an integer overflow when it returned -1 (representing
unreachability). So the method has been changed to return an `int` instead.

+++ link_state/topology/route_finder.cpp
@@ -50,7 +50,7 @@ void RouteFinder::clear_states()
     clear_queue(frontier);
 }
 
-uint16_t RouteFinder::find_next_hop(uint16_t destination_id)
+int RouteFinder::find_next_hop(uint16_t destination_id)
 {
     if (predecessors[destination_id].empty())
         return UNREACHABLE;
